Tag Maker - renamed package to align w feature package naming conventions; Added % support to LowercaseWithDashesConverterImpl
