+++ bundle/src/main/java/com/adobe/acs/tools/tag_maker/impl/TagMakerServlet.java
@@ -18,12 +18,12 @@
  * #L%
  */
 
-package com.adobe.acs.tools.tagmaker.impl;
+package com.adobe.acs.tools.tag_maker.impl;
 
 
-import com.adobe.acs.tools.tagmaker.TagData;
-import com.adobe.acs.tools.tagmaker.tagdataconverters.TagDataConverter;
-import com.adobe.acs.tools.tagmaker.tagdataconverters.impl.DefaultConverterImpl;
+import com.adobe.acs.tools.tag_maker.TagData;
+import com.adobe.acs.tools.tag_maker.tagdataconverters.TagDataConverter;
+import com.adobe.acs.tools.tag_maker.tagdataconverters.impl.DefaultConverterImpl;
 import com.day.cq.tagging.InvalidTagFormatException;
 import com.day.cq.tagging.Tag;
 import com.day.cq.tagging.TagConstants;
